Add completion for "compile-package" subcommand
The "compile-package" is not publicly documented and is mostly meant for build
tools wrapping the Gleam compiler. It should not be advertised, as it's most
likely not the subcommand users are searching for. For this reason there is no
completion for the subcommand itself; this commit only adds completion for the
arguments of the subcommand.